Introduce GUI for configuring shader & kernel settings
A proper settings screen is used now, instead of the previous
slider. On top of that, the HXE Kernel settings can now be
adjusted.

Due to the SPV3.1 legacy mode in HXE, this launcher is technically
compatible with 3.1, and thus can be distributed to the existing
users before 3.2 comes out.