Introduce dynamic declaration of the working directory
This allows the SPV3 loader to be started from, say, the start menu.

Additionally, this technically permits the SPV3 loader to be launched
from any location; however, it's still best to impose it on people to
put the files into their SPV3 installation folder.

At startup, the working directory will be set to the path declared in:

    %APPDATA%\SPV3\install.txt

Should the file not be present, the loader will use the current folder.