Apply border-less mode only when the HCE process has officially loaded
Previously, HXE would apply the hack approx. five seconds after the HCE
process was invoked. This can prove unreliable for people who run HCE on
low-end computers or on large installations.

To mitigate that, HXE will now apply the mode when the main main menu is
initiated. This is done by scanning the process memory for a specific
value that HCE declares when the game is in the main menu.