Permit the HXE kernel to invoke HCE as elevated user
The HXE kernel can optionally invoke a HCE process as an administrator.

This would rarely be needed; however, it's been reported to fix certain
crashes on SPV3.2.

By default, the elevated mode is not enabled, as it will likely irk them
and prevent them from running SPV3 on their school computers.