Add the tile manifest and images to the project
This commit introduces the visual elements manifest XML and tile PNGs to the
SPV3 loader project. Rules have been set up to compile the files to the same
target as the SPV3 & HXE executable.

When the project is compiled, the aforementioned manifest and PNG files will be
present along with the executables for convenient distribution and compatibility
with the Windows 10 start menu.