Introduce configuration for native resolution
This commit updates the UI and logic to permit the running of SPV3.2 at
the display's native resolution.

When the user ticks the Native Resolution checkbox, the resolution text
fields will be disabled.

This has been suggested by github:VoidsShadow as a follow-up to a report
by discord:Rocky.S Rogers#4655 regarding which resolution to use.